tests: move some tests from `MimeApps` to `Config` for better coverage
