+++ src/config/main_config.rs
@@ -395,6 +395,7 @@ impl MimeAppsTable {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use pretty_assertions::assert_eq;
 
     #[test]
     fn wildcard_mimes() -> Result<()> {
@@ -622,6 +623,142 @@ mod tests {
             )?);
 
         assert_eq!(config.terminal("", false)?, "wezterm start --cwd . -e");
+
+        Ok(())
+    }
+    fn test_add_handlers(config: &mut Config) -> Result<()> {
+        config.add_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("Helix.desktop".into()),
+        )?;
+
+        // Should return first added handler
+        assert_eq!(
+            config
+                .get_handler(&Mime::from_str("text/plain")?, "", false,)?
+                .to_string(),
+            "Helix.desktop"
+        );
+
+        config.add_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("nvim.desktop".into()),
+        )?;
+
+        // Should still return first added handler
+        assert_eq!(
+            config
+                .get_handler(&Mime::from_str("text/plain")?, "", false,)?
+                .to_string(),
+            "Helix.desktop"
+        );
+
+        Ok(())
+    }
+
+    fn test_remove_handlers(config: &mut Config) -> Result<()> {
+        config.remove_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("Helix.desktop".into()),
+        )?;
+
+        // With first added handler removed, second handler replaces it
+        assert_eq!(
+            config
+                .get_handler(&Mime::from_str("text/plain")?, "", false,)?
+                .to_string(),
+            "nvim.desktop"
+        );
+
+        config.remove_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("nvim.desktop".into()),
+        )?;
+
+        // Both handlers removed, should not be any left
+        assert!(config
+            .get_handler(&Mime::from_str("text/plain")?, "", false)
+            .is_err());
+
+        Ok(())
+    }
+
+    fn test_set_handlers(config: &mut Config) -> Result<()> {
+        config.set_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("Helix.desktop".into()),
+        )?;
+
+        assert_eq!(
+            config
+                .get_handler(&Mime::from_str("text/plain")?, "", false,)?
+                .to_string(),
+            "Helix.desktop"
+        );
+
+        config.set_handler(
+            &Mime::from_str("text/plain")?,
+            &DesktopHandler::assume_valid("nvim.desktop".into()),
+        )?;
+
+        // Should return second set handler because it should replace the first one
+        assert_eq!(
+            config
+                .get_handler(&Mime::from_str("text/plain")?, "", false,)?
+                .to_string(),
+            "nvim.desktop"
+        );
+
+        Ok(())
+    }
+
+    fn test_unset_handlers(config: &mut Config) -> Result<()> {
+        config.unset_handler(&Mime::from_str("text/plain")?)?;
+
+        // Handler completely unset, should not be any left
+        assert!(config
+            .get_handler(&Mime::from_str("text/plain")?, "", false)
+            .is_err());
+
+        Ok(())
+    }
+
+    #[test]
+    fn add_and_remove_handlers() -> Result<()> {
+        let mut config = Config::default();
+
+        test_add_handlers(&mut config)?;
+        test_remove_handlers(&mut config)?;
+
+        Ok(())
+    }
+
+    #[test]
+    fn set_and_unset_handlers() -> Result<()> {
+        let mut config = Config::default();
+
+        test_set_handlers(&mut config)?;
+        test_unset_handlers(&mut config)?;
+
+        Ok(())
+    }
+
+    #[test]
+    fn add_and_unset_handlers() -> Result<()> {
+        let mut config = Config::default();
+
+        test_add_handlers(&mut config)?;
+        test_unset_handlers(&mut config)?;
+
+        Ok(())
+    }
+
+    #[test]
+    fn set_and_remove_handlers() -> Result<()> {
+        let mut config = Config::default();
+
+        test_set_handlers(&mut config)?;
+        test_remove_handlers(&mut config)?;
 
         Ok(())
     }
